Made changes as per Krystal's comments
